requirements.txt
# ============================================================
# Core Vision APIs
# ============================================================
google-generativeai>=0.8.0
openai>=1.30.0
python-dotenv>=1.0.0

# ============================================================
# PDF & Image Processing
# ============================================================
PyMuPDF>=1.24.0
pdf2image>=1.17.0
Pillow>=10.0.0

# ============================================================
# Data & Statistics
# ============================================================
pandas>=2.0.0
numpy>=1.26.0
statsmodels>=0.14.0
tqdm>=4.66.0

# ============================================================
# VLM Fine-tuning (HuggingFace backend)
# Install these only if using --backend huggingface
# ============================================================
torch>=2.2.0
transformers>=4.45.0
peft>=0.12.0
trl>=0.11.0
accelerate>=0.34.0
datasets>=2.21.0
# bitsandbytes is optional (enables 4-bit quantization for lower VRAM usage)
bitsandbytes>=0.43.0
